--- src/model/cloud-runner/services/cloud-runner-namespace.ts
@@ -1,10 +1,0 @@
-import { customAlphabet } from 'nanoid';
-import CloudRunnerConstants from './cloud-runner-constants';
-
-class CloudRunnerNamespace {
-  static generateBuildName(runNumber: string | number, platform: string) {
-    const nanoid = customAlphabet(CloudRunnerConstants.alphabet, 4);
-    return `${runNumber}-${platform.toLowerCase().replace('standalone', '')}-${nanoid()}`;
-  }
-}
-export default CloudRunnerNamespace;
