+++ src/model/input-readers/generic-input-reader.ts
@@ -1,4 +1,4 @@
-import { CloudRunnerSystem } from '../cli/remote-client/remote-client-services/cloud-runner-system';
+import { CloudRunnerSystem } from '../cloud-runner/remote-client/cloud-runner-system';
 
 export class GenericInputReader {
   public static async Run(command) {
