+++ src/model/input-readers/github-cli.ts
@@ -1,4 +1,4 @@
-import { CloudRunnerSystem } from '../cli/remote-client/remote-client-services/cloud-runner-system';
+import { CloudRunnerSystem } from '../cloud-runner/remote-client/cloud-runner-system';
 import * as core from '@actions/core';
 
 export class GithubCliReader {
